refactor: extract input sidebar helpers
